+++ NekoRainbowEngine/NekoRainbowEngine/ModuleViewport.cpp
@@ -27,7 +27,7 @@ bool ModuleViewport::Start()
 	root_object = CreateGameObject("Root Object");
 	camera_test = CreateGameObject("Camera", root_object);
 	camera_test->CreateComponent(COMPONENT_CAMERA);
-	App->importer->ImportFBX("./Assets/BakerHouse.fbx");
+	App->importer->ImportFile("./Assets/BakerHouse.fbx");
 	return ret;
 }
